+++ web/src/utils/toErrorMap.ts
@@ -7,3 +7,4 @@ export const toErrorMap = (errors: FieldError[]) => {
   });
   return errorMap;
 };
+